requirements.txt
streamlit>=1.33
snowflake-connector-python>=3.6
pandas>=2.0
requests>=2.31
python-dateutil>=2.8
